Week 4/Day 1/airbnb-auditor/requirements.txt
chainlit==0.7.700
langchain==0.2.5
langchain_community==0.2.5
langchain_core==0.2.9
langchain_huggingface==0.0.3
langchain_text_splitters==0.2.1
langchain_openai==0.1.9
python-dotenv==1.0.1
pymupdf==1.24.5
qdrant-client==1.9.2